+++ actix-web-actors/src/lib.rs
@@ -1,5 +1,8 @@
-#![allow(clippy::borrow_interior_mutable_const)]
 //! Actix actors integration for Actix web framework
+
+#![deny(rust_2018_idioms)]
+#![allow(clippy::borrow_interior_mutable_const)]
+
 mod context;
 pub mod ws;