+++ CHANGES.md
@@ -3,11 +3,16 @@
 ## Unreleased - 2020-xx-xx
 ### Added
 * `middleware::NormalizePath` now has configurable behaviour for either always having a trailing
-  slash, or as the new addition, always trimming trailing slashes.
+  slash, or as the new addition, always trimming trailing slashes. [#1639]
 
 ### Changed
-* Update actix-codec and actix-utils dependencies.
+* Update actix-codec and actix-utils dependencies. [#1634]
+* `FormConfig` and `JsonConfig` configurations are now also considered when set
+  using `App::data`. [#1641]
 
+[#1639]: https://github.com/actix/actix-web/pull/1639
+[#1641]: https://github.com/actix/actix-web/pull/1641
+[#1634]: https://github.com/actix/actix-web/pull/1634
 
 ## 3.0.0-beta.3 - 2020-08-17
 ### Changed