+++ CHANGES.md
@@ -9,10 +9,14 @@
 * Update actix-codec and actix-utils dependencies. [#1634]
 * `FormConfig` and `JsonConfig` configurations are now also considered when set
   using `App::data`. [#1641]
+* `HttpServer::maxconn` is renamed to the more expressive `HttpServer::max_connections`. [#1655]
+* `HttpServer::maxconnrate` is renamed to the more expressive
+  `HttpServer::max_connection_rate`. [#1655]
 
 [#1639]: https://github.com/actix/actix-web/pull/1639
 [#1641]: https://github.com/actix/actix-web/pull/1641
 [#1634]: https://github.com/actix/actix-web/pull/1634
+[#1655]: https://github.com/actix/actix-web/pull/1655
 
 ## 3.0.0-beta.3 - 2020-08-17
 ### Changed