+++ CHANGES.md
@@ -1,6 +1,9 @@
 # Changes
 
 ## Unreleased - 2020-xx-xx
+
+
+## 3.0.0-beta.4 - 2020-09-09
 ### Added
 * `middleware::NormalizePath` now has configurable behaviour for either always having a trailing
   slash, or as the new addition, always trimming trailing slashes. [#1639]